Basic debug interface setup for geolocation info.
There's a start/end button to track when a path "starts" and "ends" (when
the buttons are pressed), and a timer and coordinates are displayed. I don't have a mobile device that can access the Internet, and my laptop isn't really cooperating when I try roaming around campus with it, so if someone could upload it to a generic server and test out if the coordinates are updating, that'd be great.

I left room for adding some results functionality, and the sending of
"path complete" data to wherever that needs to go. If there's anything
further you want me to do with this, or if I misinterpreted something, let
me know.