Apologies for the delay. Format should be better.
For some detail on what I had to swap. I needed the function to call
itself, which required I give it a self-identifier. The reason the
function needed to be able to call itself is because it needs to wait
until the document is loaded so that it can actually reference the buttons. It should still be anonymous outside of the context of the function itself, so I
don't think that will pose a significant issue.

That's the primary reason I placed the script after the body, initially.
The rest of the formatting stuff was just my being incompetent, and I
believe I fixed it according to the suggestions. I also kept the
timer/results dummy section, because I figured leaving it in would just be
simpler. We can simply remove it later if we choose not to relate them.

I'm also having some trouble with Maven, which is part of why I've waited
a while to upload what I have. If we have time I'll address it
tomorrow in class.