Make Jetty use the jarfile's path
Previously, the webserver was loading content from the current working
directory. This reworks some code so that we form an actual object so
that we can get access to the class loader, but I think it's actually a
positive move regardless.

Now, going to localhost:8080 should serve the demo.html file to you. You
can change the url by hand for now to get to our other html files.