Moved map rendering into a View
The map rendering was being done in our GpsTracker model. Models are for
data: not display. This makes a new `MapView` view and an extended
`GpsMapView` which interfaces with `GpsTracker` model.

Some niceties were added as well: an accuracy circle, adjusted zoom
levels, and auto-updating the map as you walk (not tested).