Fixes JSHint issue Colin was having
The maven plugin for JSHint we're using is rather buggy. I used JSHint
under nodejs to find the problem and fixed it, which fixes the
compilation.

This commit also includes (as an added bonus!) a configuration file for
jshint (if you want to run it from the command line yourself).